+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.5.1] - 2022-07-20
+
 ### Added
 * Added `.overrideToolchain` as a convenience for using a custom rust toolchain
 
@@ -137,6 +139,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.5.1]: https://github.com/ipetkov/crane/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/ipetkov/crane/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/ipetkov/crane/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/ipetkov/crane/compare/v0.3.3...v0.4.0